server/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
Flask-Testing==0.8.0
flake8==3.8.4
flask_restful==0.3.8
flask-bcrypt==0.7.1
coverage==5.3
flask-migrate==2.5.3
requests==2.25.0
flask-jwt-extended==3.25.0
flask-redis==0.4.0
fakeredis==1.4.5
ariadne==0.12.0